Add addV() and inject() to ScalaGraph
Similar to ScalaGraph.V() and E(), adds addV() and inject() which use
the graph's traversal source to begin a new traversal.

addV() in particular is useful when using a remote traversal source as
it is possible that a withRemote created traversal source uses an
EmptyGraph and thus addVertex() will not work.