enable arrow syntax for remote graphs by passing ScalaGraph implicitly
Prior to this one couldn't use the arrow syntax when using a remote
graph. The underlying problem is that the configuration for remote is
not stored in the graph instance, but in the TraversalSource.
Because of that, one cannot simply call e.g. `vertex.addEdge` any
more, because that doesn't know about the TraversalSource, and
therefor the remote graph.

The best way to work around that IMO is to (implicitly) pass the
TraversalSource when using the arrows DSL (e.g. vertex1 --- "label"
--> vertex2). Since IMO the graph instance should hold that
information (`ScalaGraph` does by holding onto the TraversalSource), I
decided to add that as an implicit for the arrow DSL. I.e. from now on
you need to have an implicit `ScalaGraph` in scope, then the arrow DSL
works fine with remote and local graphs.

Also: you can't specify the id of a vertex when adding it to the
graph any more. This wasn't supported by many graphs anyway, and none of the
remote graphs. It's better to let the graph allocate an ID anyway, so
I just dropped that feature.