Add limit(1) to head/headOption
This change ensures that head and headOption steps will limit
traversals to a single result, rather than executing a full traversal,
throwing away the tail, and then returning only the head.  This is
especially important when connecting to a remote GremlinServer, or
when a local Graph connects to a remote storage backend, where query
latencies can be quite large.